Move send via notify back to comms processor
